Display the forecast data using the API
getting the specific data from the forecast data retrieved, putting them in the right place. Created another function to format the dt day data into actual weekdays' names. Then used the index of the daily forecast array to limit the number of days we need displayed on the HTML page